requirements.txt
weaviate-client==3.25.3
python-dotenv==1.0.0
